Add QueryMultiStore support
