--- sc/types.go
@@ -1,5 +1,0 @@
-package sc
-
-type StateCommit interface {
-	Commit() (int64, error)
-}
